backend/requirements.txt
task-caldav-lib @ git+https://github.com/bbrendon/task-caldav-lib.git
Django==5.0.6
djangorestframework==3.15.2
django-cors-headers==4.4.0
caldav==1.3.9
icalendar==5.0.14
python-dateutil==2.9.0
requests==2.32.3
gunicorn==22.0.0
whitenoise==6.7.0
